Reorganize saved-register layout on IRQ assembly entrypoint
Previously, the registers were saved on the IRQ mode stack in an
interleaved format most efficient for single-use. But because of
ordering requirements about which registers were free at which
times, the saved registers weren't in-order:

  R15
  CPSR
  R12
  ...
  R0

This isn't a friendly orering for any kind of block-transfer to
data structures or easy re-use of code fetching/storing the live
registers among several processor modes.

This changeset reorders the stack storage format to be completely
naturally ordered:

  CPSR
  R15
  ...
  R0

This requires slightly more opcodes to accomplish, but allows the
resulting logic to be factored out into macros.

This change will be followed soon by another to implement process
termination in response to data aborts.